rename calendar to avoid name collision, fix relative imports
